// src/types/themes/themeTypes.ts
export enum CvThemes {
  DEFAULT = 'Default',
  THEME_ONE = 'Theme One',
  THEME_TWO  = 'Theme Two',
  THEME_THREE  = 'Theme Three',
  THEME_FOUR  = 'Theme Four',
  THEME_FIVE  = 'Theme Five',
  THEME_SIX = 'Theme Six',
  THEME_SEVEN = 'Theme Seven',
  THEME_EIGHT = 'Theme Eight',
  THEME_NINE = 'Theme Nine',
  THEME_TEN = 'Theme Ten'
}